Added check for default 'vcredist-x64' Juju resource
Due to licensing issues, the charm can't have a default usable resource
for Visual C++ Redistributable. At the moment, a charm with resources
cannot be published without a default resource for every resource.
This commit adds a check for the default Cloudbase Juju resource uploaded
in charm store for 'vcredist-x64', in order to error out in case users
rely on that default one.